pyproject.toml
[project]
name = "app-shadowing-practice"
version = "0.1.0"
description = "English shadowing practice app for individual use"
authors = [
    { name = "Your Name", email = "hogehoge@example.com" }
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "openai==1.35.9",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "pydantic>=2.5.0",
    "jinja2>=3.1.2",
    "aiosqlite>=0.19.0",
    "greenlet>=3.1.1",
]
readme = "README.md"
requires-python = "==3.10.12"

[project.scripts]
dev = "src.app:start_dev_server"
prod = "src.app:start_prod_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.8",
    "mypy>=1.11.2",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]


[tool.ruff]
lint.select = [
    "F", # Flake8
    "B", # Black
    "I", # isort
    "E", # error
    "W"  # warning
]
lint.ignore = []
lint.fixable = ["ALL"]
exclude = [
    ".ruff_cache"
]
line-length = 119
indent-width = 4
target-version = "py310"

[[tool.mypy.overrides]]
module = [
    'sklearn.*',
    'google.cloud.*',
    'google.cloud.storage.*',
    'tqdm.*',
    'plotly.*',
    'dotenv.*',
    'ipdb.*',
    'statsmodels.*',
    'tests.*',
]
ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning"
]
addopts = "-s"
